requirements.txt
flask==3.0.3
werkzeug==3.0.3
pdfminer.six==20240706
python-docx==1.1.2
scikit-learn==1.5.2
numpy==2.1.2
scipy==1.13.1


